code/requirements.txt
aiogram>=3.3.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
neo4j>=5.15.0
apscheduler>=3.10.4